Tests: Reorganize validator tests
